Add reusable grow animation
The exploding golden tooth animation is now implemented using this
functionality
